make the driver code implementation-independent
All of the DS402Driver code is actually independent of DS402 (all
the specificities are in DS402Channel). Make the code independent
through a ChannelBase interface.
